fix ingredient type checks
